+++ ZenWare.Loader/resource.h
@@ -8,6 +8,6 @@
 // заголовком окна, сплэшем и VERSIONINFO exe автоматически.
 #define ZENWARE_VER_MAJOR 3
 #define ZENWARE_VER_MINOR 8
-#define ZENWARE_VER_PATCH 1
+#define ZENWARE_VER_PATCH 2
 
-#define ZENWARE_VER_STR "3.8.1"
+#define ZENWARE_VER_STR "3.8.2"